backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
pytest
httpx
python-multipart
python-dotenv
qrcode[pil]
